added module to package json
