+++ test/integration.test.cjs
@@ -3,8 +3,8 @@ const expect = require("chai").expect;
 const app = require("../server/app.cjs");
 
 describe("Integration test", function () {
-  it("should return ok", async function () {
-    const response = await request(app).get("/healthcheck");
-    expect(response.text).equal("ok");
-  });
+    it("should return ok", async function () {
+        const response = await request(app).get("/healthcheck");
+        expect(response.text).equal("ok");
+    });
 });
